Ejercicio 08 del Lab de .Net
Se le agregaron validaciones, y endpoints especificos para consultar la API

+++ Lab.Ejercicio004.EF.Logic/CategoriesLogic.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using Lab.Ejercicio004.EF.Entities;
+using Lab.Ejercicio004.EF.Utils;
 
 namespace Lab.Ejercicio004.EF.Logic
 {
@@ -11,12 +12,19 @@ namespace Lab.Ejercicio004.EF.Logic
         {
             try
             {
-                _context.Categories.Add(itemToAdd);
-                _context.SaveChanges();
+                if (CategoriesValidation.ValidateCategory(itemToAdd.CategoryName, itemToAdd.Description))
+                {
+                    _context.Categories.Add(itemToAdd);
+                    _context.SaveChanges();
+                }
+                else
+                {
+                    return "There are errors in the data provided";
+                }
             }
-            catch (Exception ex)
+            catch (Exception)
             {
-                return $"Error to insert Data into Categories {ex.Message}";
+                return $"Error to insert Data into Categories";
             }
             return "Data successfully inserted";
         }
@@ -26,23 +34,38 @@ namespace Lab.Ejercicio004.EF.Logic
             try
             {
                 Categories categoryToDelete = _context.Categories.Find(id);
-                _context.Categories.Remove(categoryToDelete);
-                _context.SaveChanges();
+                if (categoryToDelete == null)
+                {
+                    return $"The category you want to delete does not exist {id}";
+                }
+                else
+                {
+                    _context.Categories.Remove(categoryToDelete);
+                    _context.SaveChanges();
+                }
             }
-            catch (System.Data.Entity.Infrastructure.DbUpdateException updateException)
+            catch (System.Data.Entity.Infrastructure.DbUpdateException)
             {
-                return updateException.InnerException.InnerException.Message;
+                return "The category you want to delete is related to another table";
             }
-            catch (Exception ex)
+            catch (Exception)
             {
-                return $"Error to delete Category {ex.Message} - {ex.GetType()}";
+                return "Unknown error detected when trying to delete a category";
             }
-            return "Category Deleted Successfully";
+            return "Category successfully deleted";
         }
 
         public override List<Categories> GetAll()
         {
-            return _context.Categories.ToList();
+            try
+            {
+                return _context.Categories.ToList();
+            }
+            catch (Exception)
+            {
+                return null;
+            }
+
         }
 
         public override Categories GetEntityByID(int id)
@@ -63,13 +86,24 @@ namespace Lab.Ejercicio004.EF.Logic
         {
             try
             {
-                var categoryToUpdate = _context.Categories.Find(itemToChange.CategoryID);
-                _context.Entry(categoryToUpdate).CurrentValues.SetValues(itemToChange);
-                _context.SaveChanges();
+                if (CategoriesValidation.ValidateCategory(itemToChange.CategoryName, itemToChange.Description))
+                {
+                    var categoryToUpdate = _context.Categories.Find(itemToChange.CategoryID);
+                    if (categoryToUpdate == null)
+                    {
+                        return $"The category you want to update does not exist {itemToChange.CategoryID}";
+                    }
+                    _context.Entry(categoryToUpdate).CurrentValues.SetValues(itemToChange);
+                    _context.SaveChanges();
+                }
+                else
+                {
+                    return "There are errors in the data provided";
+                }
             }
-            catch (Exception ex)
+            catch (Exception)
             {
-                return $"Error to update category details {ex.Message}";
+                return "Unknown error detected when trying to update category";
             }
             return "All data modify successfully";
         }
