+++ Lab.Ejercicio004.EF.Logic/ProductsLogic.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using Lab.Ejercicio004.EF.Entities;
+using Lab.Ejercicio004.EF.Utils;
 
 namespace Lab.Ejercicio004.EF.Logic
 {
@@ -11,12 +12,19 @@ namespace Lab.Ejercicio004.EF.Logic
         {
             try
             {
-                _context.Products.Add(itemToAdd);
-                _context.SaveChanges();
+                if (ProductsValidation.ValidateProduct(itemToAdd.ProductName,itemToAdd.QuantityPerUnit,itemToAdd.UnitPrice.ToString()))
+                {
+                    _context.Products.Add(itemToAdd);
+                    _context.SaveChanges();
+                }
+                else
+                {
+                    return "There are errors in the data provided";
+                }
             }
-            catch (Exception ex)
+            catch (Exception)
             {
-                return $"Error to insert Data into Products {ex.Message}";
+                return $"Error to insert Data into Products";
             }
             return "Data successfully inserted";
         }
@@ -26,36 +34,79 @@ namespace Lab.Ejercicio004.EF.Logic
             try
             {
                 Products productToDelete = _context.Products.Find(id);
-                _context.Products.Remove(productToDelete);
-                _context.SaveChanges();
+                if (productToDelete == null)
+                {
+                    return $"The product you want to delete does not exist {id}";
+                }
+                else
+                {
+                    _context.Products.Remove(productToDelete);
+                    _context.SaveChanges();
+                }
             }
-            catch (System.Data.Entity.Infrastructure.DbUpdateException updateException)
+            catch (System.Data.Entity.Infrastructure.DbUpdateException)
             {
-                return updateException.InnerException.InnerException.Message;
+                return "The product you want to delete is related to another table";
             }
-            catch (Exception ex)
+            catch (Exception)
             {
-                return $"Error to delete Product {ex.Message} - {ex.GetType()}";
+                return "Unknown error detected when trying to delete a Product";
             }
             return "Product Deleted Successfully";
         }
 
         public override List<Products> GetAll()
         {
-            return _context.Products.ToList();
+            try
+            {
+                return _context.Products.ToList();
+            }
+            catch (Exception)
+            {
+                return null;
+            }
+
+        }
+
+        public override Products GetEntityByID(int id)
+        {
+            Products oProduct;
+            try
+            {
+                oProduct = _context.Products.Find(id);
+                return oProduct;
+            }
+            catch (Exception)
+            {
+            }
+            return null;
         }
 
         public override string Update(Products itemToChange)
         {
             try
             {
-                var productsToUpdate = _context.Products.Find(itemToChange.ProductID);
-                _context.Entry(productsToUpdate).CurrentValues.SetValues(itemToChange);
-                _context.SaveChanges();
+                if (ProductsValidation.ValidateProduct(itemToChange.ProductName, itemToChange.QuantityPerUnit, itemToChange.UnitPrice.ToString()))
+                {
+                    var productsToUpdate = _context.Products.Find(itemToChange.ProductID);
+                    if (productsToUpdate == null)
+                    {
+                        return $"The product you want to update does not exist {itemToChange.CategoryID}";
+                    }
+                    else
+                    {
+                        _context.Entry(productsToUpdate).CurrentValues.SetValues(itemToChange);
+                        _context.SaveChanges();
+                    }
+                }
+                else
+                {
+                    return "There are errors in the data provided";
+                }
             }
-            catch (Exception ex)
+            catch (Exception)
             {
-                return $"Error to update product {ex.Message}";
+                return "Unknown error detected when trying to update product";
             }
             return "All data modify successfully";
         }
@@ -74,9 +125,9 @@ namespace Lab.Ejercicio004.EF.Logic
                     }
                 }
             }
-            catch (Exception ex)
+            catch (Exception)
             {
-                return $"Error to set to null to categories ID {ex.Message}";
+                return $"Error to set to null to categories ID";
             }
             return $"The selected category id: {id} is successfully set to null";
         }
@@ -95,25 +146,13 @@ namespace Lab.Ejercicio004.EF.Logic
                     }
                 }
             }
-            catch (Exception ex)
+            catch (Exception)
             {
-                return $"Error to set to null to supplier ID {ex.Message}";
+                return $"Error to set to null to supplier ID";
             }
             return $"The selected supplier id: {id} is successfully set to null";
         }
 
-        public override Products GetEntityByID(int id)
-        {
-            Products oProduct;
-            try
-            {
-                oProduct = _context.Products.Find(id);
-                return oProduct;
-            }
-            catch (Exception)
-            {
-            }
-            return null;
-        }
+
     }
 }
